add get endpoint

+++ src/main/java/edu/mit/dos/persistence/DigitalObjectJpaRepository.java
@@ -2,14 +2,11 @@ package edu.mit.dos.persistence;
 
 import edu.mit.dos.model.DigitalObject;
 import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.data.jpa.repository.Query;
-//import org.springframework.data.repository.query.Param;
-
 
 
 public interface DigitalObjectJpaRepository extends JpaRepository<DigitalObject, Integer> {
 
-    DigitalObject findByOid(int oid);
+    DigitalObject findByOid(long oid);
 
     DigitalObject findByHandle(String handle);
